Refactor: main.dart와 app.dart 분리 및 Firebase 초기화 코드 추가
